+++ Sources/Harmonic/AppDelegate.swift
@@ -7,5 +7,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationDidFinishLaunching(_ notification: Notification) {
         NSApp.setActivationPolicy(.accessory)
         statusBarController = StatusBarController()
+        UpdateService.shared.checkOnLaunchIfNeeded()
     }
 }